% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{cat0}
\alias{cat0}
\title{cat0}
\usage{
cat0(...)
}
\arguments{
\item{...}{R objects.}
}
\description{
\code{cat0} is an alternative to \link{cat}
}
\details{
This is indeed the same as \code{cat(..., sep = '')}
}
\examples{
cat0('string', 'withoutspaces')

}
