% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{catn}
\alias{catn}
\title{catn}
\usage{
catn(...)
}
\arguments{
\item{...}{R objects.}
}
\description{
\code{catn} is an alternative to \link{cat}
}
\details{
This is indeed the same as \code{\link{cat0}} but with a new line
}
\examples{
catn('string', 'withoutspaces', 'andanewline')

}
