% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{last}
\alias{last}
\title{last}
\usage{
last(x)
}
\arguments{
\item{x}{data.frame or array}
}
\description{
\code{last} is based on \link{tail}
}
\details{
It returns the last element of an array or dataframe
}
\examples{
last(c(1,2,3))
}
