% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{progress}
\alias{progress}
\title{progress}
\usage{
progress(x, max = 100)
}
\arguments{
\item{x}{current position}

\item{max}{max range}
}
\description{
\code{progress} is an original function from \url{https://github.com/klmr/rcane/blob/1bb57f8a4340b6ac618f98fb9f81964caea9ab8c/system.R#L24}
}
\details{
It is an alternative function to \link{txtProgressBar}
and \link{setTxtProgressBar}
}
\examples{
progress(25)
}
