% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{stop0}
\alias{stop0}
\title{stop0}
\usage{
stop0()
}
\description{
\code{stop0} is an alternative to \link{stop}
}
\details{
This is indeed the same as \code{\link{stop}} but quietly, very
useful when included in the middle of a source file.
}
\examples{
stop0()
}
