% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{library0}
\alias{library0}
\title{library0}
\usage{
library0(package)
}
\arguments{
\item{package}{the name of the package, only given as a name.}
}
\description{
\code{library0} is an alternative to \link{library}, which also installs the package if needed.
}
\details{
This is an alternative function to \link{library}, which pretends to be an easier way
to attach and install a package in the same step. You can indicate the package as a string or
directly by the name.
}
